+++ apps/prototype/src/pages/child/PhraseBuilderPage.tsx
@@ -75,7 +75,7 @@ export const PhraseBuilderPage: React.FC = () => {
 
   return (
     <div className="flex flex-col min-h-[calc(100vh-80px)]">
-      <ChildTopBar showSettings={false} />
+      <ChildTopBar />
 
       {/* Header: back + clear */}
       <div className="flex items-center gap-2.5 px-5 pt-1 pb-0.5">